# frontend/src/api/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "EntitlementRead": {
        "description": "Ce à quoi une organisation a droit, et ce qu'elle a consommé.\n\nC'est la réponse qui alimente la page compte **et** les boîtes de dialogue des murs de\npaiement : une seule route, pour que les deux ne puissent pas afficher deux vérités.",
        "properties": {
          "archived_project_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Archived Project Ids",
            "type": "array"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "integer"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanRead"
          },
          "subscription": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubscriptionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "trial_available": {
            "title": "Trial Available",
            "type": "boolean"
          },
          "trial_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Ends At"
          },
          "usage": {
            "items": {
              "$ref": "#/components/schemas/UsageRead"
            },
            "title": "Usage",
            "type": "array"
          }
        },
        "required": [
          "organization_id",
          "plan",
          "subscription",
          "period_start",
          "period_end",
          "trial_available",
          "trial_ends_at",
          "usage",
          "archived_project_ids"
        ],
        "title": "EntitlementRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanRead": {
        "description": "Un palier de la grille, tel qu'il est en base.\n\n`limits` et `features` sortent bruts : la page tarifs les affiche avec les libellés servis à\ncôté, et n'en code aucun en dur. Une limite déplacée par `UPDATE` change donc la page sans\ndéploiement, ce qui est exactement l'intention de `plan_catalog`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "features": {
            "additionalProperties": true,
            "title": "Features",
            "type": "object"
          },
          "limits": {
            "additionalProperties": true,
            "title": "Limits",
            "type": "object"
          },
          "monthly_price_cents": {
            "title": "Monthly Price Cents",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "seat_price_cents": {
            "title": "Seat Price Cents",
            "type": "integer"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "tagline": {
            "title": "Tagline",
            "type": "string"
          },
          "yearly_price_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yearly Price Cents"
          }
        },
        "required": [
          "code",
          "name",
          "tagline",
          "monthly_price_cents",
          "yearly_price_cents",
          "seat_price_cents",
          "currency",
          "limits",
          "features",
          "sort_order"
        ],
        "title": "PlanRead",
        "type": "object"
      },
...
      "SubscriptionRead": {
        "properties": {
          "cancel_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel At"
          },
          "current_period_end": {
            "format": "date-time",
            "title": "Current Period End",
            "type": "string"
          },
          "current_period_start": {
            "format": "date-time",
            "title": "Current Period Start",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "plan_code": {
            "title": "Plan Code",
            "type": "string"
          },
          "seats": {
            "title": "Seats",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/SubscriptionStatus"
          },
          "trial_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Ends At"
          }
        },
        "required": [
          "id",
          "plan_code",
          "status",
          "current_period_start",
          "current_period_end",
          "trial_ends_at",
          "cancel_at",
          "seats"
        ],
        "title": "SubscriptionRead",
        "type": "object"
      },
      "SubscriptionStatus": {
        "description": "Cycle de vie d'un abonnement.\n\nFermé à quatre valeurs, contrairement aux codes de plan : ce sont les états que le code sait\ninterpréter, et en inventer un cinquième par `INSERT` rendrait une organisation ni autorisée\nni refusée. C'est donc bien un ENUM en base.\n\n`past_due` reste **autorisé** : couper l'accès d'un client dont le prélèvement a échoué le\nmatin même, en plein chantier, coûte plus cher que les quelques jours de relance. La coupure\nest un passage explicite en `canceled`.",
        "enum": [
          "trialing",
          "active",
          "past_due",
          "canceled"
        ],
        "title": "SubscriptionStatus",
        "type": "string"
      },
...
      "UsageRead": {
        "description": "Consommation d'une métrique sur la période de facturation en cours.\n\n`limit` à `null` veut dire **illimité**, jamais zéro : les confondre afficherait « 3 / 0 » à un\nclient qui n'a aucun plafond.",
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "metric",
          "value",
          "limit"
        ],
        "title": "UsageRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Éditeur de plan de rénovation 2D → 3D",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/organizations/{organization_id}/subscription": {
      "get": {
        "description": "Droits, abonnement et consommation de l'organisation.\n\nC'est ici, et seulement ici, que le **déclassement** est appliqué : les chantiers excédentaires\npassent en lecture seule (`archived_at`), jamais à la corbeille. Le faire pendant une simple\nlecture de projet ferait écrire n'importe quelle requête ; le faire dans une tâche planifiée\ndemanderait un ordonnanceur que le dépôt n'a pas. La page compte et la boîte de dialogue du\nmur de paiement passent toutes les deux par ici, donc la réconciliation a lieu au moment exact\noù l'utilisateur regarde ce qu'il lui reste.",
        "operationId": "read_subscription_api_organizations__organization_id__subscription_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Subscription",
        "tags": [
          "abonnement"
        ]
      }
    },
    "/api/organizations/{organization_id}/subscription/trial": {
      "post": {
        "description": "Ouvre l'essai de 14 jours du palier Artisan, sans carte.\n\nIdempotente en pratique : un essai déjà consommé ne rend pas une erreur mais l'état courant.\nRenvoyer un 409 obligerait le frontend à traiter un cas qui ne change rien pour l'utilisateur —\nil voulait connaître ses droits, il les obtient.\n\n`editor` suffit : c'est le rôle qui pose le geste monétisé (dessiner, chiffrer, exporter), et\nl'essai est déclenché par ce geste-là. Exiger `admin` ferait échouer l'ouverture automatique au\nmoment précis où elle doit être invisible.",
        "operationId": "open_trial_api_organizations__organization_id__subscription_trial_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Open Trial",
        "tags": [
          "abonnement"
        ]
      }
    },
...
  }
}